backend/requirements.txt
fastapi
pydantic
uvicorn[standard]
newspaper3k
lxml_html_clean
trafilatura
